+++ src/MinEdLauncher/Process.fs
@@ -9,7 +9,7 @@ let launchProcesses (processes:ProcessStartInfo list) =
             Process.Start(p) |> Some
         with
         | e ->
-            Log.exn e $"Unable to start pre-launch process %s{p.FileName}"
+            Log.exn e $"Unable to start process %s{p.FileName}"
             None)
 
 let stopProcesses (processes: Process list) =
@@ -22,3 +22,15 @@ let stopProcesses (processes: Process list) =
             p.StandardError.ReadToEnd() |> ignore
             Log.info $"Stopped process %s{p.ProcessName}"
         | Error msg -> Log.warn msg)
+    
+let writeOutput (processes: Process list) =
+    processes
+    |> List.iter(fun p ->
+        use p = p
+        p.EnableRaisingEvents <- true
+        p.OutputDataReceived.Add(fun a -> if a.Data <> null then printfn $"  %s{a.Data}")
+        p.ErrorDataReceived.Add(fun a -> if a.Data <> null then printfn $"  %s{a.Data}")
+        p.BeginErrorReadLine()
+        p.BeginOutputReadLine()
+        p.WaitForExit()
+    )